Se cambio el nombre de la clase Tile a Tablero, se agregaron comentarios a todas las clases, y se edito el codigo para facilitar la lectura.
