+++ paniqueados2/Clases/Trazo.cs
@@ -3,80 +3,65 @@ using System;
 using System.Text.RegularExpressions;
 
 namespace Clases { 
-
+    // Clase Trazo, que representa el recorrido que ha tenido un Jugador hasta un punto
     public class Trazo
     {
-        string path;
-        int inicioX;
-        int inicioY;
-        int tamanio;
+        //// ATRIBUTOS
+            // Camino que ha recorrido el Jugador, en una cadena con los inputs
+            string path;
 
-        public Trazo(int tam)
-        {
-            this.path = "X";
-            this.tamanio = tam;
-        }
+            // Coordenadas del punto inicial del Trazo
+            int inicioX;
+            int inicioY;
 
-        public int getX()
-        {
-            return this.inicioX;
-        }
+            // Tamaño del sprite del trazo (es tanto largo como ancho)
+            int tamanio;
 
-        public int getY()
-        {
-            return this.inicioY;
-        }
-
-        public char backtrack()
-        {
-            int length = this.path.Length - 1;
-            char res = this.path[length];
-            if (res != 'X') { this.path = this.path.Remove(length); }
-            return (res);
-        }
-
-        public void nuevaDireccion(string dir)
-        {
-            this.path += dir;
-            if (this.path.Length > 1)
+        //// MÉTODOS
+            // Constructor
+            public Trazo(int tam)
             {
-                this.borrarRedundantes();
+                this.path = "X";
+                this.tamanio = tam;
             }
-        }
+            
+            // Métodos GET para obtener los atributos de la instancia
+            public int getX() { return this.inicioX; }
+            public int getY() { return this.inicioY; }
+            public int getTam() { return this.tamanio; }
+            public string getPath() { return this.path; }
+            
+            // Métodos SET para asignar valores a los atributos de la instancia
+            public void setXInicial(int xNueva) { this.inicioX = xNueva; }
+            public void setYInicial(int yNueva) { this.inicioY = yNueva; }
+            public void setPath(string nuevo) { this.path = nuevo; }
 
-        public void borrarRedundantes()
-        {
-            string cadena1 = this.path;
-            string patronRegex1 = "WS|SW";
-            string patronRegex2 = "AD|DA";
-            string vacio = "";
-            cadena1 = Regex.Replace(cadena1, patronRegex1, vacio);
-            string cadenaSinRedundancia = Regex.Replace(cadena1, patronRegex2, vacio);
-            this.path = cadenaSinRedundancia;
-        }
+            // Métodos ESPECÍFICOS de la instancia
 
-        public string getPath()
-        {
-            return this.path;
-        }
-
-        public void setXInicial(int xNueva)
-        {
-            this.inicioX = xNueva;
-        }
-
-        public void setYInicial(int yNueva)
-        {
-            this.inicioY = yNueva;
-        }
-
-        public int getTam() {
-            return this.tamanio;
-        }
-
-        public void setPath(string nuevo) {
-            this.path = nuevo;
-        }
+                    // Regresa un paso hacia atrás, borrando él último movimiento registrado en el camino
+                    public char backtrack() {
+                        int length = this.path.Length - 1;
+                        char res = this.path[length];
+                        if (res != 'X') this.path = this.path.Remove(length); // El caracter X representa el inicio del camino.
+                        return (res);
+                    }
+            
+                    // Agrega una nueva dirección a la cadena del camino.
+                    public void nuevaDireccion(string dir) {
+                        this.path += dir;
+                        if (this.path.Length > 1) this.borrarRedundantes();
+                    }
+            
+                    // Elimina los inputs que se cancelen entre sí de la cadena con propósitos de acortarla y poder regresar borrando el exceso.
+                    public void borrarRedundantes() {
+                        string cadena1 = this.path;
+                        // Regex de los conjuntos de caracteres redundantes
+                        string patronRegex1 = "WS|SW";
+                        string patronRegex2 = "AD|DA";
+                        cadena1 = Regex.Replace(cadena1, patronRegex1, "");
+                        string cadenaSinRedundancia = Regex.Replace(cadena1, patronRegex2, "");
+                        this.path = cadenaSinRedundancia;
+                    }
 
     }
